feat: update adding and deleting enabled payment methods

+++ auth-service/src/main/java/com/psp/authservice/service/UserService.java
@@ -5,6 +5,7 @@ import com.psp.authservice.model.EnabledPaymentMethod;
 import com.psp.authservice.model.PaymentMethod;
 import com.psp.authservice.model.RegularUser;
 import com.psp.authservice.model.User;
+import com.psp.authservice.repository.EnabledPaymentMethodRepository;
 import com.psp.authservice.repository.RegularUserRepository;
 import com.psp.authservice.repository.UserRepository;
 import com.psp.authservice.security.util.TokenUtils;
@@ -25,6 +26,8 @@ import java.util.UUID;
 @Service
 @Slf4j
 public class UserService implements UserDetailsService {
+    @Autowired
+    private EnabledPaymentMethodRepository enabledPaymentMethodRepository;
 
     @Autowired
     private UserRepository userRepository;
@@ -96,39 +99,46 @@ public class UserService implements UserDetailsService {
         PaymentMethod paymentMethod = paymentMethodService.findPaymentMethodById(enabledPaymentMethodDto.getPaymentMethod().getId());
         if (getPaymentMethodForCompany(user, paymentMethod.getId()) == null) {
             enabledPaymentMethod.setPaymentMethod(paymentMethod);
-            enablePaymentMethodForCompany(user, enabledPaymentMethod);
+            List<EnabledPaymentMethod> enabledPaymentMethods = enablePaymentMethodForCompany(user, enabledPaymentMethod);
             log.debug("Payment method with id: {}, enabled for merchant: {}", paymentMethod.getId(), user.getId());
-            return new ResponseEntity<>(HttpStatus.CREATED);
+            return new ResponseEntity<>(enabledPaymentMethods.stream().map(enabledPayment -> modelMapper.map(enabledPayment, EnabledPaymentMethodDto.class)),HttpStatus.CREATED);
         } else {
             log.warn("Payment method with id: {} is already enabled for merchant: {}", paymentMethod.getId(), user.getId());
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
 
-    private void enablePaymentMethodForCompany(RegularUser user, EnabledPaymentMethod enabledPaymentMethod) {
+    private List<EnabledPaymentMethod> enablePaymentMethodForCompany(RegularUser user, EnabledPaymentMethod enabledPaymentMethod) {
         enabledPaymentMethod = enabledPaymentMethodService.save(enabledPaymentMethod);
         List<EnabledPaymentMethod> userEnabledPaymentMethods = user.getEnabledPaymentMethods();
         userEnabledPaymentMethods.add(enabledPaymentMethod);
         user.setEnabledPaymentMethods(userEnabledPaymentMethods);
         userRepository.save(user);
+        return user.getEnabledPaymentMethods();
     }
 
     public ResponseEntity<?> deleteEnabledPaymentMethod(String userEmail, UUID enabledPaymentMethodId) {
         RegularUser user = (RegularUser) userRepository.findByEmail(userEmail);
         if (isPaymentMethodEnabledForCompany(user, enabledPaymentMethodId)) {
-            deleteEnabledPaymentMethod(enabledPaymentMethodId, user);
+            List<EnabledPaymentMethod> enabledPaymentMethods = deleteEnabledPaymentMethod(enabledPaymentMethodId, user);
             log.debug("Enabled payment method option with id: {}, deleted from merchant: {}", enabledPaymentMethodId, user.getId());
-            return new ResponseEntity<>(HttpStatus.OK);
+            return new ResponseEntity<>(enabledPaymentMethods.stream().map(enabledPaymentMethod -> modelMapper.map(enabledPaymentMethod, EnabledPaymentMethodDto.class)),HttpStatus.OK);
         }
         log.debug("Merchant {} cannot delete payment method option with id: {} - no payment option with given id", user.getId(), enabledPaymentMethodId);
         return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-    private void deleteEnabledPaymentMethod(UUID enabledPaymentMethodId, RegularUser user) {
+    private List<EnabledPaymentMethod> deleteEnabledPaymentMethod(UUID enabledPaymentMethodId, RegularUser user) {
         EnabledPaymentMethod enabledPaymentMethod = enabledPaymentMethodService.findById(enabledPaymentMethodId);
-        user.getEnabledPaymentMethods().remove(enabledPaymentMethod);
-        userRepository.save(user);
-        enabledPaymentMethodService.delete(enabledPaymentMethod);
+        for(int i =0 ; i < user.getEnabledPaymentMethods().size() ; i++) {
+            if(enabledPaymentMethod.getId().equals(user.getEnabledPaymentMethods().get(i).getId())){
+                user.getEnabledPaymentMethods().remove(user.getEnabledPaymentMethods().get(i));
+                userRepository.save(user);
+                enabledPaymentMethodService.delete(enabledPaymentMethod);
+                return user.getEnabledPaymentMethods();
+            }
+        }
+        return null;
     }
 
     private boolean isPaymentMethodEnabledForCompany(RegularUser user, UUID enabledPaymentMethodId) {
